requirements.txt
flask==3.0.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
requests==2.31.0
gunicorn==21.2.0
werkzeug>=3.0.0
loguru==0.7.2
